significantly speed up render calls (eliminate memory leak?)
I noticed that some function call was leaking memory. After looking into it
in the Performance panel of Google Chrome, I realized I was making new callback
allocation on every render for the Default callback of UniformShouldUpdate callback.

This allocation was also causing Uuid to be called for each new callback created.
In the end, I got rid of the allocation by manually specifying true as the default
rather than creating an entire new function just to do that.